+++ cost_map_ros/CHANGELOG.rst
@@ -2,5 +2,15 @@
 Changelog
 =========
 
+0.3.1 (2017-02-16)
+------------------
+* minor api name fixes for Costmap2D references
+
+0.3.0 (2016-12-27)
+------------------
+* costmap2DROS converters
+* bugfix missing data copy in the toGridMap conversion
+
 0.2.0 (2016-08-25)
 ------------------
+* initial creation of similar functionality to that of grid maps